Copy script to local: - Added xdebug configuration for ngixn

+++ Docker/php-fpm/scripts/php_extension_install.sh
@@ -3,11 +3,15 @@
 # add wget
 apt-get update -yqq && apt-get -f install -yyq wget
 
+FILE=/tmp/scripts/install-php-extensions.sh
+if ! [ -f "$FILE" ]; then
 # download helper script
 # @see https://github.com/mlocati/docker-php-extension-installer/
 wget -q -O /usr/local/bin/install-php-extensions https://raw.githubusercontent.com/mlocati/docker-php-extension-installer/master/install-php-extensions \
     || (echo "Failed while downloading php extension installer!"; exit 1)
-
+else
+  cp $FILE /usr/local/bin/install-php-extensions
+fi
 # install extensions
 chmod uga+x /usr/local/bin/install-php-extensions && sync && install-php-extensions \
     opcache \
@@ -17,5 +21,4 @@ chmod uga+x /usr/local/bin/install-php-extensions && sync && install-php-extensi
     bz2 \
     bcmath \
     redis \
-    zip \
 ;
